refactor: 전략 패턴에 사용되는 테스트용 클래스 분리
